Rename texts file
